rag_backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi",
    "pillow>=12.0.0",
    "pydantic",
    "pypdf>=6.1.3",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "supabase",
    "uvicorn>=0.38.0",
]
